Fix incorrect Git clone URL in readme.md
The organisation name and repository name were the wrong name round.

Also switch from HTTP to HTTPS, which is more secure (GitHub redirects
to HTTPS if one tries to use HTTP).

Also insert a hyphen in the destination directory's name to match the
repository's name (which is the default destination directory).