Switch from squidguard to a nginx reverseproxy for more flexibility
Also upgrade squid, do MITM so ssl sites still get a block page, update
the block page to be a little bit better (since it's served by squid now
and not squidguard).

Net effect should be basically the same, but it's now possible to
whitelist individual pages on a domain, rather than the whole domain.